refactored DeviceDao

+++ src/main/java/org/buyandfly/dao/IDao.java
@@ -3,7 +3,7 @@ package org.buyandfly.dao;
 import java.util.List;
 
 public interface IDao<T> {
-    T getById(final int id);
+    T getById(final long id);
     List<T> getAll();
     void add(final T t);
     void edit(final T t);
